requirements.txt
django~=3.2.7
djangorestframework==3.13.1
django-filter==22.1
graphene_django
django-graphql-jwt
gunicorn==20.1.0
whitenoise==5.3.0
dj-database-url==0.5.0

